app/requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aiozeroconf==0.1.8
async-timeout==3.0.1
attrs==19.3.0
blinker==1.4
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
curve25519-donna==1.3
ecdsa==0.15
ed25519==1.5
Flask==1.1.1
h11==0.9.0
h2==3.1.1
hpack==3.0.0
Hypercorn==0.9.0
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
multidict==4.7.4
nest-asyncio==1.2.3
netifaces==0.10.9
priority==1.3.0
protobuf==3.11.2
pyatv==0.4.0a12
pycparser==2.19
PyYAML==5.3
Quart==0.10.0
six==1.14.0
sortedcontainers==2.1.0
srptools==1.0.0
tlslite-ng==0.7.5
toml==0.10.0
typing-extensions==3.7.4.1
Werkzeug==0.16.0
wsproto==0.15.0
yarl==1.4.2
